Add Benefit model validations
Because: I want a benefit name to be present and unique scoped to a category,
and because I want a benefit category to be present and selected from a
list of benefit categories

This commit: Adds benefit model validations for presence to both name
and category. Name also has a unique case-insensitive constraint scoped
to category. Category must be included in a list of the categories array
which is a constant set on the benefit model