Benefit feature (#17)
* Feature: Benefit Model

Because: I want to represent international health insurance plan
benefits

This commit: Creates a benefit model with a name and category

* Add Benefit model validations

Because: I want a benefit name to be present and unique scoped to a category,
and because I want a benefit category to be present and selected from a
list of benefit categories

This commit: Adds benefit model validations for presence to both name
and category. Name also has a unique case-insensitive constraint scoped
to category. Category must be included in a list of the categories array
which is a constant set on the benefit model