Feature: Product Modules
Because: I want to represent health insurance product modules

This commit: Adds a ProductModule model. The model belongs_to a product
and has a name, category and sum assured attributes

A model is unique if the name is unique scoped to the category. All
fields are required to be valid