Product module feature (#18)
* Feature: Product Modules

Because: I want to represent health insurance product modules

This commit: Adds a ProductModule model. The model belongs_to a product
and has a name, category and sum assured attributes

A model is unique if the name is unique scoped to the category. All
fields are required to be valid

* ProductModule validations

Because: I want the ProductModule to validate attributes

This commit: Makes the following validation
* Validates presence of name
* Validates uniqueness of name scoped to category
* validates presence of category
* validates inclusion of category in a CATEGORIES array constant
* validates presence of sum_assured
* validates sum_assured is in a valid format