Install stimulusJS
Stimulus will be used to handle the parts of the app that require some
JS functionality.

Initially it will be used to handle selecting health insurance products
on the new comparison view. The products select will need to be updated
everytime a new insurer is selected. Once a product is chosen the
product modules will also need to be updated