Add Products to Product Select
Because: I want to add an insurers products to the product select box
when an insurer is selected on the new comparison page

This commit: Sends a remote fetch request to the products controller
sending the insurer id.

The request is sent for a json response and the controller provides a
json response of the selected insurers products. The json response is
then converted into an array and iterated creating a new option for the
product select box and appending it.

When an insurer is selected the product select is cleared