Add LinkedProductModule model
Because: Only those product modules linked with the core product module
should be shown on the comparison product form for the product module
radio buttons

This commit: Introduces a LinkedProductModule model. The model linked to
the core product module and a linked_module which itself is another
product module. The idea is to link all "sub_modules" to a core module
so that when a core module is selected on the product module select, it
can then show only those linked product modules