Refactor the comparison product view
Because: The current view meant controllers with too many instance
variables

This commit: Refactors the view to move some of the instance variables
to view helpers.

This also reduces the complexity involved in getting the correct
benefits to be displayed in the show excel view as currently there is
only one option which can affect how the benefits are displayed. This
meant there was too little complexity to require a query object and a
service object just to handle one paramater option